Healthbar and healthscript is checked

+++ TE4Pixadome/Assets/Harry/Scripts/Healtbar.cs
@@ -7,12 +7,16 @@ public class Healtbar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxHealth(int health)
+    //sätter maxvärdet slidern kan  har och hur lång den nuvarande är
+    public void SetMaxHealth(float health)
     {
         slider.maxValue = health;
         slider.value = health;
     }
-    public void SetHealthValue(int health)
+
+    //sätter det nuvaranade värdet till slidern 
+    //används så att slidern faktisk ändrar storlek
+    public void SetHealthValue(float health)
     {
         slider.value = health;
     }
